+++ src/app/fyle/my-view-report/my-view-report.page.ts
@@ -8,7 +8,9 @@ import { map, tap, switchMap, finalize, shareReplay } from 'rxjs/operators';
 import { TransactionService } from 'src/app/core/services/transaction.service';
 import { AuthService } from 'src/app/core/services/auth.service';
 import { LoaderService } from 'src/app/core/services/loader.service';
-import { AlertController } from '@ionic/angular';
+import { AlertController, PopoverController } from '@ionic/angular';
+import { PopupService } from 'src/app/core/services/popup.service';
+import { ShareReportComponent } from './share-report/share-report.component';
 
 @Component({
   selector: 'app-my-view-report',
@@ -36,7 +38,8 @@ export class MyViewReportPage implements OnInit {
     private authService: AuthService,
     private loaderService: LoaderService,
     private router: Router,
-    private alertController: AlertController
+    private popupService: PopupService,
+    private popoverController: PopoverController
   ) { }
 
   ngOnInit() {
@@ -129,44 +132,42 @@ export class MyViewReportPage implements OnInit {
   }
 
   async deleteReport() {
-    const message = `
-    <p class="highlight-info">
-    On deleting this report, all the associated expenses will be moved to <strong>"My Expenses"</strong> list.</p>
-    <p>Are you sure, you want to delete this report?</p>`;
-    const alert = await this.alertController.create({
+    const popupResults = await this.popupService.showPopup({
       header: 'Delete Report',
-      message,
-      buttons: [
-        {
-          text: 'Cancel',
-          role: 'cancel',
-          cssClass: 'secondary',
-          handler: noop
-        }, {
-          text: 'Okay',
-          handler: () => {
-            from(this.loaderService.showLoader()).pipe(
-              switchMap(() => {
-                return this.reportService.delete(this.activatedRoute.snapshot.params.id)
-              }),
-              finalize(() => from(this.loaderService.hideLoader()))
-            ).subscribe(() => {
-              this.router.navigate(['/', 'enterprise', 'my_reports']);
-            })
-          }
-        }
-      ]
+      message: `
+      <p class="highlight-info">
+        On deleting this report, all the associated expenses will be moved to <strong>My Expenses</strong> list.
+      </p>
+      <p>
+        Are you sure, you want to delete this report?
+      </p>
+      `,
+      primaryCta: {
+        text: 'Okay'
+      },
+      secondaryCta: {
+        text: 'Cancel'
+      }
     });
 
-    await alert.present();
+    if (popupResults === 'primary') {
+      from(this.loaderService.showLoader()).pipe(
+        switchMap(() => {
+          return this.reportService.delete(this.activatedRoute.snapshot.params.id)
+        }),
+        finalize(() => from(this.loaderService.hideLoader()))
+      ).subscribe(() => {
+        this.router.navigate(['/', 'enterprise', 'my_reports']);
+      })
+    }
   }
 
   showResubmitReportSummaryPopover() {
-
+    // TODO
   }
 
   showSubmitReportSummaryPopover() {
-
+    // TODO
   }
 
   goToTransaction(etxn: any) {
@@ -178,9 +179,13 @@ export class MyViewReportPage implements OnInit {
     }
 
     if (category === 'activity') {
-      // showCannotEditActivityDialog();
-      alert('can not edit Activity -> TODO show can not edit activity Dialog');
-      return;
+      this.popupService.showPopup({
+        header: 'Cannot Edit Activity',
+        message: 'Editing activity is not supported in mobile app.',
+        primaryCta: {
+          text: 'Cancel'
+        }
+      });
     }
 
     let route;
@@ -207,39 +212,25 @@ export class MyViewReportPage implements OnInit {
   }
 
   async shareReport(event) {
-    const alert = await this.alertController.create({
-      header: 'Share Report',
-      message: 'Share report via email',
-      inputs: [
-        {
-          name: 'email',
-          type: 'email',
-          placeholder: 'Email ID'
-        }
-      ],
-      buttons: [
-        {
-          text: 'Cancel',
-          role: 'cancel',
-          cssClass: 'secondary',
-          handler: noop
-        }, {
-          text: 'Okay',
-          handler: (e) => {
-            const data = {
-              report_ids: [this.activatedRoute.snapshot.params.id],
-              email: e.email
-            };
-            this.reportService.downloadSummaryPdfUrl(data).subscribe(async () => {
-              const message = `We will send ${e.email} a link to download the PDF <br> when it is generated and send you a copy.`;
-              await this.loaderService.showLoader(message);
-            });
-          }
-        }
-      ]
+    const popover = await this.popoverController.create({
+      component: ShareReportComponent,
+      cssClass: 'dialog-popover'
     });
 
-    await alert.present();
+    await popover.present();
+
+    const { data } = await popover.onWillDismiss();
+
+    if (data.email) {
+      const params = {
+        report_ids: [this.activatedRoute.snapshot.params.id],
+        email: data.email
+      };
+      this.reportService.downloadSummaryPdfUrl(params).subscribe(async () => {
+        const message = `We will send ${data.email} a link to download the PDF <br> when it is generated and send you a copy.`;
+        await this.loaderService.showLoader(message);
+      });
+    }
   }
 
   canEditTxn(txState) {
